Extract span info creation logic into _create_span_info utility
- Add _create_span_info() helper function to encapsulate span metadata conversion
- Replace inline span conversion loop with clean list comprehension
- Better separation of concerns and code reusability
- All tests pass with refactored implementation
