+++ mlflow/genai/judges/tools/__init__.py
@@ -1,12 +1,13 @@
 from mlflow.genai.judges.tools.base import JudgeTool
 from mlflow.genai.judges.tools.get_trace_info import GetTraceInfoTool
-from mlflow.genai.judges.tools.list_spans import ListSpansResult, ListSpansTool, SpanInfo
+from mlflow.genai.judges.tools.list_spans import ListSpansResult, ListSpansTool
 from mlflow.genai.judges.tools.registry import (
     JudgeToolRegistry,
     invoke_judge_tool,
     list_judge_tools,
     register_judge_tool,
 )
+from mlflow.genai.judges.tools.types import SpanInfo
 
 __all__ = [
     "JudgeTool",
