+++ tests/genai/judges/test_judge_tool_list_spans.py
@@ -11,8 +11,8 @@ from mlflow.entities.trace_data import TraceData
 from mlflow.entities.trace_info import TraceInfo
 from mlflow.entities.trace_location import TraceLocation
 from mlflow.entities.trace_state import TraceState
-from mlflow.genai.judges.tools.constants import ToolNames
-from mlflow.genai.judges.tools.list_spans import ListSpansResult, ListSpansTool, SpanInfo
+from mlflow.genai.judges.tools.list_spans import ListSpansResult, ListSpansTool
+from mlflow.genai.judges.tools.types import SpanInfo
 
 
 @pytest.fixture
@@ -49,25 +49,6 @@ def mock_trace_with_spans():
     return Trace(info=trace_info, data=trace_data)
 
 
-def test_list_spans_tool_name():
-    """Test that the tool returns the correct name."""
-    tool = ListSpansTool()
-    assert tool.name == ToolNames.LIST_SPANS
-
-
-def test_list_spans_tool_get_definition():
-    """Test that the tool returns a valid definition."""
-    tool = ListSpansTool()
-    definition = tool.get_definition()
-
-    assert definition.function.name == "list_spans"
-    assert "List information about spans within a trace" in definition.function.description
-    assert definition.function.parameters.type == "object"
-    assert "max_results" in definition.function.parameters.properties
-    assert "page_token" in definition.function.parameters.properties
-    assert definition.function.parameters.required == []
-
-
 def test_list_spans_tool_invoke_success(mock_trace_with_spans):
     """Test that the tool successfully returns span data."""
     tool = ListSpansTool()
@@ -98,19 +79,29 @@ def test_list_spans_tool_invoke_success(mock_trace_with_spans):
     assert span2.duration_ms == 300.0  # 0.3 seconds
 
 
-def test_list_spans_tool_invoke_empty_trace():
-    """Test that the tool handles traces with no spans."""
-    trace_info = TraceInfo(
-        trace_id="empty-trace",
-        trace_location=TraceLocation.from_experiment_id("0"),
-        request_time=1234567890,
-        state=TraceState.OK,
-    )
-    trace_data = TraceData(request="{}", response="{}", spans=[])
-    empty_trace = Trace(info=trace_info, data=trace_data)
-
+@pytest.mark.parametrize(
+    ("trace_input", "test_name"),
+    [
+        (None, "none_trace"),
+        ("empty_trace", "empty_trace"),
+    ],
+)
+def test_list_spans_tool_invoke_empty_cases(trace_input, test_name):
+    """Test that the tool handles empty cases (None trace or no spans)."""
     tool = ListSpansTool()
-    result = tool.invoke(empty_trace)
+
+    # Handle the empty trace case by creating it inline
+    if trace_input == "empty_trace":
+        trace_info = TraceInfo(
+            trace_id="empty-trace",
+            trace_location=TraceLocation.from_experiment_id("0"),
+            request_time=1234567890,
+            state=TraceState.OK,
+        )
+        trace_data = TraceData(request="{}", response="{}", spans=[])
+        trace_input = Trace(info=trace_info, data=trace_data)
+
+    result = tool.invoke(trace_input)
 
     assert isinstance(result, ListSpansResult)
     assert len(result.spans) == 0
@@ -151,13 +142,3 @@ def test_list_spans_tool_invoke_invalid_page_token(mock_trace_with_spans):
         MlflowException, match="Invalid page_token '\\[\\]': must be a valid integer"
     ):
         tool.invoke(mock_trace_with_spans, page_token=[])
-
-
-def test_list_spans_tool_invoke_none_trace():
-    """Test that the tool handles None trace gracefully."""
-    tool = ListSpansTool()
-    result = tool.invoke(None)
-
-    assert isinstance(result, ListSpansResult)
-    assert len(result.spans) == 0
-    assert result.next_page_token is None
